requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==4.0.4
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2021.3
sqlparse==0.4.2
